#2 - code refactoring
