--- src/actions.ts
@@ -1,4 +1,0 @@
-export enum Actions {
-  GetMetadataAndContent = 'getMetadataAndContent',
-  Block = 'Block',
-}
